refactor: extract draggable signature component
