+++ frontend/src/pages/DocumentWorkflow.js
@@ -9,6 +9,7 @@ import logService from '../services/logService';
 import sanitize from '../utils/sanitize';
 import Countdown from '../components/Countdown';
 import useIsMobile from '../hooks/useIsMobile';
+import DraggableSignature from '../components/DraggableSignature';
 
 pdfjs.GlobalWorkerOptions.workerSrc = '/pdf.worker.min.js';
 
@@ -19,238 +20,6 @@ function debounce(func, wait) {
     timeout = setTimeout(() => func(...args), wait);
   };
 }
-
-/* =========================
-   COMPOSANT SIGNATURE DRAGGABLE (conservé)
-   ========================= */
-const DraggableSignature = React.memo(function DraggableSignature({
-  field,
-  pageWidth,
-  pageHeight,
-  isMobileView,
-  onUpdate,
-  onDelete
-}) {
-  const [isDragging, setIsDragging] = useState(false);
-  const [isResizing, setIsResizing] = useState(false);
-  const [dragStart, setDragStart] = useState({ x: 0, y: 0, fieldX: 0, fieldY: 0 });
-  const [resizeStart, setResizeStart] = useState({ x: 0, y: 0, width: 0, height: 0 });
-  const elementRef = useRef(null);
-
-  const style = useMemo(() => ({
-    position: 'absolute',
-    left: field.position.x * pageWidth,
-    top: field.position.y * pageHeight,
-    width: field.position.width * pageWidth,
-    height: field.position.height * pageHeight,
-    borderRadius: 8,
-    boxShadow: '0 0 0 1px rgba(0,0,0,.20), 0 2px 6px rgba(0,0,0,.08)',
-    background: '#fff',
-    display: 'flex',
-    alignItems: 'center',
-    justifyContent: 'center',
-    zIndex: 15,
-    cursor: isDragging ? 'grabbing' : 'grab',
-    border: '2px solid transparent',
-    userSelect: 'none',
-  }), [field.position, pageWidth, pageHeight, isDragging]);
-
-  const handleMouseDown = useCallback((e) => {
-    if (e.target.classList.contains('resize-handle')) return;
-    e.preventDefault();
-    e.stopPropagation();
-    setIsDragging(true);
-    setDragStart({
-      x: e.clientX,
-      y: e.clientY,
-      fieldX: field.position.x,
-      fieldY: field.position.y
-    });
-  }, [field.position]);
-
-  const handleTouchStart = useCallback((e) => {
-    if (e.target.classList.contains('resize-handle')) return;
-    const touch = e.touches[0];
-    if (!touch) return;
-    e.preventDefault();
-    e.stopPropagation();
-    setIsDragging(true);
-    setDragStart({
-      x: touch.clientX,
-      y: touch.clientY,
-      fieldX: field.position.x,
-      fieldY: field.position.y,
-    });
-  }, [field.position]);
-
-  const handleResizeStart = useCallback((e) => {
-    e.preventDefault();
-    e.stopPropagation();
-    setIsResizing(true);
-    setResizeStart({
-      x: e.clientX,
-      y: e.clientY,
-      width: field.position.width,
-      height: field.position.height
-    });
-  }, [field.position]);
-
-  const handleResizeTouchStart = useCallback((e) => {
-    const touch = e.touches[0];
-    if (!touch) return;
-    e.preventDefault();
-    e.stopPropagation();
-    setIsResizing(true);
-    setResizeStart({
-      x: touch.clientX,
-      y: touch.clientY,
-      width: field.position.width,
-      height: field.position.height,
-    });
-  }, [field.position]);
-
-  useEffect(() => {
-    if (!isDragging && !isResizing) return;
-
-    const handleMouseMove = (e) => {
-      if (isDragging) {
-        const deltaX = (e.clientX - dragStart.x) / pageWidth;
-        const deltaY = (e.clientY - dragStart.y) / pageHeight;
-        const newPosition = {
-          ...field.position,
-          x: Math.max(0, dragStart.fieldX + deltaX),
-          y: Math.max(0, dragStart.fieldY + deltaY)
-        };
-        onUpdate(field, { position: newPosition });
-      } else if (isResizing) {
-        const deltaX = (e.clientX - resizeStart.x) / pageWidth;
-        const deltaY = (e.clientY - resizeStart.y) / pageHeight;
-        const newPosition = {
-          ...field.position,
-          width: Math.max(50 / pageWidth, resizeStart.width + deltaX),
-          height: Math.max(20 / pageHeight, resizeStart.height + deltaY)
-        };
-        onUpdate(field, { position: newPosition });
-      }
-    };
-
-    const handleMouseUp = () => {
-      setIsDragging(false);
-      setIsResizing(false);
-    };
-
-    const handleTouchMove = (e) => {
-      const touch = e.touches[0];
-      if (!touch) return;
-      e.preventDefault();
-      if (isDragging) {
-        const deltaX = (touch.clientX - dragStart.x) / pageWidth;
-        const deltaY = (touch.clientY - dragStart.y) / pageHeight;
-        const newPosition = {
-          ...field.position,
-          x: Math.max(0, dragStart.fieldX + deltaX),
-          y: Math.max(0, dragStart.fieldY + deltaY),
-        };
-        onUpdate(field, { position: newPosition });
-      } else if (isResizing) {
-        const deltaX = (touch.clientX - resizeStart.x) / pageWidth;
-        const deltaY = (touch.clientY - resizeStart.y) / pageHeight;
-        const newPosition = {
-          ...field.position,
-          width: Math.max(50 / pageWidth, resizeStart.width + deltaX),
-          height: Math.max(20 / pageHeight, resizeStart.height + deltaY),
-        };
-        onUpdate(field, { position: newPosition });
-      }
-    };
-
-    const handleTouchEnd = () => {
-      setIsDragging(false);
-      setIsResizing(false);
-    };
-
-    document.addEventListener('mousemove', handleMouseMove);
-    document.addEventListener('mouseup', handleMouseUp);
-    document.addEventListener('touchmove', handleTouchMove, { passive: false });
-    document.addEventListener('touchend', handleTouchEnd);
-
-    return () => {
-      document.removeEventListener('mousemove', handleMouseMove);
-      document.removeEventListener('mouseup', handleMouseUp);
-       document.removeEventListener('touchmove', handleTouchMove);
-       document.removeEventListener('touchend', handleTouchEnd);
-    };
-  }, [isDragging, isResizing, dragStart, resizeStart, pageWidth, pageHeight, field, onUpdate]);
-
-  return (
-    <div
-      ref={elementRef}
-      style={style}
-      onMouseDown={handleMouseDown}
-      onTouchStart={handleTouchStart}
-      onTouchMove={(e) => e.preventDefault()}
-      onTouchEnd={(e) => e.preventDefault()}
-      className={`signature-field ${isDragging ? 'dragging' : ''} hover:shadow-lg transition-shadow`}
-    >
-      <div
-        style={{
-          textAlign: 'center',
-          fontSize: isMobileView ? 10 : 12,
-          lineHeight: 1.1,
-          color: '#374151',
-          pointerEvents: 'none'
-        }}
-      >
-        <div style={{ fontWeight: 700, marginBottom: 2 }}>Signature</div>
-        <div>{field.recipient_name || field.name.replace('Signature ', '')}</div>
-      </div>
-
-      <div
-        className="resize-handle"
-        onMouseDown={handleResizeStart}
-        onTouchStart={handleResizeTouchStart}
-        style={{
-          position: 'absolute',
-          bottom: -4,
-          right: -4,
-          width: 12,
-          height: 12,
-          background: '#3b82f6',
-          borderRadius: '50%',
-          cursor: 'se-resize',
-          border: '2px solid white',
-          boxShadow: '0 2px 4px rgba(0,0,0,0.2)'
-        }}
-      />
-
-      <button
-        onClick={(e) => {
-          e.stopPropagation();
-          onDelete(field);
-        }}
-        style={{
-          position: 'absolute',
-          top: -8,
-          right: -8,
-          width: 20,
-          height: 20,
-          background: '#ef4444',
-          borderRadius: '50%',
-          border: '2px solid white',
-          color: 'white',
-          fontSize: 12,
-          cursor: 'pointer',
-          display: 'flex',
-          alignItems: 'center',
-          justifyContent: 'center',
-          boxShadow: '0 2px 4px rgba(0,0,0,0.2)'
-        }}
-      >
-        ×
-      </button>
-    </div>
-  );
-});
 
 /* =========================
    PDF VIEWER
@@ -334,6 +103,7 @@ const PDFViewer = React.memo(function PDFViewer({
                               pageWidth={pageWidthPx}
                               pageHeight={pageHeightPx}
                               isMobileView={isMobileView}
+                              tapToPlace={isMobileView}
                               onUpdate={onUpdateField}
                               onDelete={onDeleteField}
                             />
